feat(:js) Ruby2JS to Manual Rollup + Propshaft
Ruby2JS just relies too much on bundlers for Lit
to properly work. Instead of keeping this around,
write proper JS for Lit Elements and utilize the
Lit approved rollup config to generate a single
build file that Propshaft can utilize.

This and even the folks behind the Ruby2JS project
confirmed that the future of frontend tooling in Ruby
is likely WASM and not Ruby2JS. Until there is a
way to go from WASM -> WebComponents, let's abandon
this approach.